requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
RPi.GPIO>=0.7.0
torch>=2.0.0